TP03: HTTP y manejo de errores implementados en DetalleReceta

+++ src/main/java/com/imb2025/smedico/service/jpa/DetalleRecetaServicelmpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class DetalleRecetaServicelmpl implements IDetalleRecetaService {
@@ -19,18 +20,40 @@ public class DetalleRecetaServicelmpl implements IDetalleRecetaService {
         return repository.findAll();
     }
 
+    
     @Override
     public DetalleReceta findById(Long id) {
-        return repository.findById(id).orElse(null);
+        return repository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("DetalleReceta no encontrado con id: " + id));
+    }
+
+
+    @Override
+    public boolean existsById(Long id) {
+        return repository.existsById(id);
     }
 
     @Override
-    public DetalleReceta save(DetalleReceta recetaMedicamento) {
-        return repository.save(recetaMedicamento);
+    public DetalleReceta create(DetalleReceta detalleReceta) {
+        return repository.save(detalleReceta);
     }
 
+ 
+    @Override
+    public DetalleReceta update(Long id, DetalleReceta detalleReceta) {
+        if (!repository.existsById(id)) {
+            throw new NoSuchElementException("No se puede actualizar. DetalleReceta no encontrado con id: " + id);
+        }
+        detalleReceta.setId(id);
+        return repository.save(detalleReceta);
+    }
+
+    
     @Override
     public void deleteById(Long id) {
+        if (!repository.existsById(id)) {
+            throw new NoSuchElementException("No se puede eliminar. DetalleReceta no encontrado con id: " + id);
+        }
         repository.deleteById(id);
     }
 }
